refactor: combine packages using vite-express
