--- frontend/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "aepp-faucet",
-  "private": true,
-  "version": "1.0.0",
-  "description": "Faucet Aepp",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
-    "preview": "vite preview"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/aeternity/aepp-faucet-nodejs.git"
-  },
-  "keywords": [
-    "blockchain",
-    "aeternity",
-    "faucet"
-  ],
-  "author": "Enrico Icardi <enrico@welance.com>",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/aeternity/aepp-faucet-nodejs/issues"
-  },
-  "homepage": "https://github.com/aeternity/aepp-faucet-nodejs#readme",
-  "devDependencies": {
-    "typescript": "^5.2.2",
-    "vite": "^5.2.0"
-  }
-}
