+++ src/result.ts
